Refactor task editing logic: streamline file deletion checks and ensure proper handling of uploaded files in editSingleTask function.

+++ controllers/adminController.js
@@ -622,11 +622,10 @@ const editSingleTask = async (req, res) => {
         fileName: uploadedFileUrl.original_filename,
       };
       task.attachments = [newAttachment]; // Assuming only one attachment is allowed
+      if (fs.existsSync(req.file.path)) fs.unlinkSync(req.file.path);
     }
 
     await task.save();
-
-    if (fs.existsSync(req.file.path)) fs.unlinkSync(req.file.path);
 
     return res
       .status(200)
@@ -637,7 +636,9 @@ const editSingleTask = async (req, res) => {
       );
   } catch (err) {
     console.error("Error updating task:", err);
-    if (fs.existsSync(req.file.path)) fs.unlinkSync(req.file.path);
+    if (req.file) {
+      if (fs.existsSync(req.file.path)) fs.unlinkSync(req.file.path);
+    }
     return res
       .status(500)
       .redirect(
